Move external_identifier_to_url_mapper to services
It's in the services layer not data access
