+++ wikidatarefisland/data_access/__init__.py
@@ -1,6 +1,4 @@
 from .storage import Storage
 from .wdqs_reader import WdqsReader
-from .external_identifier_to_url_mapper import (ExternalIdentifierToUrlMapper,
-                                                WdqsExternalIdentifierToUrlMapper)
 
-__all__ = [Storage, WdqsReader, ExternalIdentifierToUrlMapper, WdqsExternalIdentifierToUrlMapper]
+__all__ = [Storage, WdqsReader]
