+++ wikidatarefisland/services/__init__.py
@@ -1,0 +1,4 @@
+from .external_identifier_to_url_mapper import (ExternalIdentifierToUrlMapper,
+                                                WdqsExternalIdentifierToUrlMapper)
+
+__all__ = [ExternalIdentifierToUrlMapper, WdqsExternalIdentifierToUrlMapper]
